ref: move get_top_5_issues_by_count_for_file and get_projects_and_filenames_from_source_file
