+++ src/sentry/integrations/source_code_management/commit_context.py
@@ -4,13 +4,26 @@ import logging
 from abc import ABC, abstractmethod
 from collections.abc import Sequence
 from dataclasses import dataclass
-from datetime import datetime, timedelta, timezone
-from typing import Any
+from datetime import UTC, datetime, timedelta, timezone
+from typing import Any, TypedDict
 
 import sentry_sdk
 from django.db import connection
+from django.db.models import Value
+from django.db.models.functions import StrIndex
 from django.utils import timezone as django_timezone
-from snuba_sdk import Column, Condition, Direction, Entity, Function, Op, OrderBy, Query
+from snuba_sdk import (
+    BooleanCondition,
+    BooleanOp,
+    Column,
+    Condition,
+    Direction,
+    Entity,
+    Function,
+    Op,
+    OrderBy,
+    Query,
+)
 from snuba_sdk import Request as SnubaRequest
 
 from sentry import analytics
@@ -19,6 +32,8 @@ from sentry.constants import ObjectStatus
 from sentry.integrations.gitlab.constants import GITLAB_CLOUD_BASE_URL
 from sentry.integrations.models.repository_project_path_config import RepositoryProjectPathConfig
 from sentry.integrations.services.integration import integration_service
+from sentry.integrations.source_code_management.constants import STACKFRAME_COUNT
+from sentry.integrations.source_code_management.language_parsers import PATCH_PARSERS
 from sentry.integrations.source_code_management.metrics import (
     CommitContextHaltReason,
     CommitContextIntegrationInteractionEvent,
@@ -27,7 +42,7 @@ from sentry.integrations.source_code_management.metrics import (
 from sentry.integrations.types import ExternalProviderEnum
 from sentry.locks import locks
 from sentry.models.commit import Commit
-from sentry.models.group import Group
+from sentry.models.group import Group, GroupStatus
 from sentry.models.groupowner import GroupOwner
 from sentry.models.options.organization_option import OrganizationOption
 from sentry.models.organization import Organization
@@ -83,13 +98,24 @@ def _pr_comment_log(integration_name: str, suffix: str) -> str:
     return f"{integration_name}.pr_comment.{suffix}"
 
 
+def _open_pr_comment_log(integration_name: str, suffix: str) -> str:
+    return f"{integration_name}.open_pr_comment.{suffix}"
+
+
 PR_COMMENT_TASK_TTL = timedelta(minutes=5).total_seconds()
 PR_COMMENT_WINDOW = 14  # days
 
-MAX_SUSPECT_COMMITS = 1000
+PR_MAX_SUSPECT_COMMITS = 1000
+
+OPEN_PR_MAX_RECENT_ISSUES = 5000
+# Caps the number of files that can be modified in a PR to leave a comment
+OPEN_PR_MAX_FILES_CHANGED = 7
+# Caps the number of lines that can be modified in a PR to leave a comment
+OPEN_PR_MAX_LINES_CHANGED = 500
 
 # Metrics
 MERGED_PR_METRICS_BASE = "{integration}.pr_comment.{key}"
+OPEN_PR_METRICS_BASE = "{integration}.open_pr_comment.{key}"
 
 
 @dataclass
@@ -113,6 +139,35 @@ class CommitInfo:
 @dataclass
 class FileBlameInfo(SourceLineInfo):
     commit: CommitInfo
+
+
+@dataclass
+class PullRequestIssue:
+    title: str
+    subtitle: str | None
+    url: str
+    affected_users: int | None = None
+    event_count: int | None = None
+    function_name: str | None = None
+
+
+@dataclass
+class PullRequestFile:
+    filename: str
+    patch: str
+
+
+class IssueCount(TypedDict):
+    group_id: int
+    event_count: int
+
+
+class IssueCountForFile(TypedDict):
+    title: str
+    culprit: str
+    function_name: str
+    group_id: int
+    event_count: int
 
 
 class CommitContextIntegration(ABC):
@@ -202,6 +257,118 @@ class CommitContextIntegration(ABC):
         Given a list of source files and line numbers,returns the commit info for the most recent commit.
         """
         return self.get_blame_for_files(files, extra)
+
+    def get_projects_and_filenames_from_source_file(
+        self,
+        organization: Organization,
+        repo: Repository,
+        pr_filename: str,
+    ) -> tuple[set[Project], set[str]]:
+        # fetch the code mappings in which the source_root is a substring at the start of pr_filename
+        code_mappings = (
+            RepositoryProjectPathConfig.objects.filter(
+                organization_id=organization.id,
+                repository_id=repo.id,
+            )
+            .annotate(substring_match=StrIndex(Value(pr_filename), "source_root"))
+            .filter(substring_match=1)
+        )
+
+        project_list: set[Project] = set()
+        sentry_filenames = set()
+
+        if len(code_mappings):
+            for code_mapping in code_mappings:
+                project_list.add(code_mapping.project)
+                sentry_filenames.add(
+                    pr_filename.replace(code_mapping.source_root, code_mapping.stack_root, 1)
+                )
+        return project_list, sentry_filenames
+
+    def create_or_update_comment(
+        self,
+        repo: Repository,
+        pr_key: str,
+        comment_data: dict[str, Any],
+        pullrequest_id: int,
+        issue_ids: list[int],
+        metrics_base: str,
+        comment_type: int = CommentType.MERGED_PR,
+        language: str | None = None,
+    ):
+        client = self.get_client()
+
+        pr_comment = PullRequestComment.objects.filter(
+            pull_request__id=pullrequest_id, comment_type=comment_type
+        ).first()
+
+        interaction_type = (
+            SCMIntegrationInteractionType.CREATE_COMMENT
+            if not pr_comment
+            else SCMIntegrationInteractionType.UPDATE_COMMENT
+        )
+
+        with CommitContextIntegrationInteractionEvent(
+            interaction_type=interaction_type,
+            provider_key=self.integration_name,
+            repository=repo,
+            pull_request_id=pullrequest_id,
+        ).capture():
+            if pr_comment is None:
+                resp = client.create_comment(
+                    repo=repo.name,
+                    issue_id=str(pr_key),
+                    data=comment_data,
+                )
+
+                current_time = django_timezone.now()
+                comment = PullRequestComment.objects.create(
+                    external_id=resp.body["id"],
+                    pull_request_id=pullrequest_id,
+                    created_at=current_time,
+                    updated_at=current_time,
+                    group_ids=issue_ids,
+                    comment_type=comment_type,
+                )
+                metrics.incr(
+                    metrics_base.format(integration=self.integration_name, key="comment_created")
+                )
+
+                if comment_type == CommentType.OPEN_PR:
+                    analytics.record(
+                        "open_pr_comment.created",
+                        comment_id=comment.id,
+                        org_id=repo.organization_id,
+                        pr_id=pullrequest_id,
+                        language=(language or "not found"),
+                    )
+            else:
+                resp = client.update_comment(
+                    repo=repo.name,
+                    issue_id=str(pr_key),
+                    comment_id=pr_comment.external_id,
+                    data=comment_data,
+                )
+                metrics.incr(
+                    metrics_base.format(integration=self.integration_name, key="comment_updated")
+                )
+                pr_comment.updated_at = django_timezone.now()
+                pr_comment.group_ids = issue_ids
+                pr_comment.save()
+
+            logger_event = metrics_base.format(
+                integration=self.integration_name, key="create_or_update_comment"
+            )
+            logger.info(
+                logger_event,
+                extra={"new_comment": pr_comment is None, "pr_key": pr_key, "repo": repo.name},
+            )
+
+    @abstractmethod
+    def on_create_or_update_comment_error(self, api_error: ApiError, metrics_base: str) -> bool:
+        raise NotImplementedError
+
+    # PR Comment Workflow
 
     def queue_comment_task_if_needed(
         self,
@@ -343,89 +510,6 @@ class CommitContextIntegration(ABC):
     def queue_comment_task(self, pullrequest_id: int, project_id: int) -> None:
         raise NotImplementedError
 
-    def create_or_update_comment(
-        self,
-        repo: Repository,
-        pr_key: str,
-        comment_data: dict[str, Any],
-        pullrequest_id: int,
-        issue_ids: list[int],
-        metrics_base: str,
-        comment_type: int = CommentType.MERGED_PR,
-        language: str | None = None,
-    ):
-        client = self.get_client()
-
-        pr_comment = PullRequestComment.objects.filter(
-            pull_request__id=pullrequest_id, comment_type=comment_type
-        ).first()
-
-        interaction_type = (
-            SCMIntegrationInteractionType.CREATE_COMMENT
-            if not pr_comment
-            else SCMIntegrationInteractionType.UPDATE_COMMENT
-        )
-
-        with CommitContextIntegrationInteractionEvent(
-            interaction_type=interaction_type,
-            provider_key=self.integration_name,
-            repository=repo,
-            pull_request_id=pullrequest_id,
-        ).capture():
-            if pr_comment is None:
-                resp = client.create_comment(
-                    repo=repo.name,
-                    issue_id=str(pr_key),
-                    data=comment_data,
-                )
-
-                current_time = django_timezone.now()
-                comment = PullRequestComment.objects.create(
-                    external_id=resp.body["id"],
-                    pull_request_id=pullrequest_id,
-                    created_at=current_time,
-                    updated_at=current_time,
-                    group_ids=issue_ids,
-                    comment_type=comment_type,
-                )
-                metrics.incr(
-                    metrics_base.format(integration=self.integration_name, key="comment_created")
-                )
-
-                if comment_type == CommentType.OPEN_PR:
-                    analytics.record(
-                        "open_pr_comment.created",
-                        comment_id=comment.id,
-                        org_id=repo.organization_id,
-                        pr_id=pullrequest_id,
-                        language=(language or "not found"),
-                    )
-            else:
-                resp = client.update_comment(
-                    repo=repo.name,
-                    issue_id=str(pr_key),
-                    comment_id=pr_comment.external_id,
-                    data=comment_data,
-                )
-                metrics.incr(
-                    metrics_base.format(integration=self.integration_name, key="comment_updated")
-                )
-                pr_comment.updated_at = django_timezone.now()
-                pr_comment.group_ids = issue_ids
-                pr_comment.save()
-
-            logger_event = metrics_base.format(
-                integration=self.integration_name, key="create_or_update_comment"
-            )
-            logger.info(
-                logger_event,
-                extra={"new_comment": pr_comment is None, "pr_key": pr_key, "repo": repo.name},
-            )
-
-    @abstractmethod
-    def on_create_or_update_comment_error(self, api_error: ApiError, metrics_base: str) -> bool:
-        raise NotImplementedError
-
     @abstractmethod
     def format_pr_comment(self, issue_ids: list[int]) -> str:
         raise NotImplementedError
@@ -441,7 +525,9 @@ class CommitContextIntegration(ABC):
     ) -> dict[str, Any]:
         raise NotImplementedError
 
-    def get_issue_ids_from_pr(self, pr: PullRequest, limit: int = MAX_SUSPECT_COMMITS) -> list[int]:
+    def get_issue_ids_from_pr(
+        self, pr: PullRequest, limit: int = PR_MAX_SUSPECT_COMMITS
+    ) -> list[int]:
         with connection.cursor() as cursor:
             cursor.execute(
                 """
@@ -458,9 +544,7 @@ class CommitContextIntegration(ABC):
             )
             return [issue_id for (issue_id,) in cursor.fetchall()]
 
-    def get_top_5_issues_by_count(
-        self, issue_ids: list[int], project: Project
-    ) -> list[dict[str, Any]]:
+    def get_top_5_issues_by_count(self, issue_ids: list[int], project: Project) -> list[IssueCount]:
         """Given a list of issue group ids, return a sublist of the top 5 ordered by event count"""
         request = SnubaRequest(
             dataset=Dataset.Events.value,
@@ -492,7 +576,7 @@ class CommitContextIntegration(ABC):
         cache_key = debounce_pr_comment_cache_key(pullrequest_id=pr.id)
 
         # cap to 1000 issues in which the merge commit is the suspect commit
-        issue_ids = self.get_issue_ids_from_pr(pr, limit=MAX_SUSPECT_COMMITS)
+        issue_ids = self.get_issue_ids_from_pr(pr, limit=PR_MAX_SUSPECT_COMMITS)
 
         if not OrganizationOption.objects.get_value(
             organization=organization,
@@ -568,6 +652,178 @@ class CommitContextIntegration(ABC):
                 tags={"type": "api_error"},
             )
             raise
+
+    # Open PR Comment Workflow
+
+    def get_top_5_issues_by_count_for_file(
+        self, projects: list[Project], sentry_filenames: list[str], function_names: list[str]
+    ) -> list[IssueCountForFile]:
+        """
+        Given a list of projects, filenames reverse-codemapped into filenames in Sentry,
+        and function names representing the list of functions changed in a PR file, return a
+        sublist of the top 5 recent unhandled issues ordered by event count.
+        """
+        if not len(projects):
+            return []
+
+        patch_parsers = PATCH_PARSERS
+        # NOTE: if we are testing beta patch parsers, add check here
+
+        # fetches the appropriate parser for formatting the snuba query given the file extension
+        # the extension is never replaced in reverse codemapping
+        language_parser = patch_parsers.get(sentry_filenames[0].split(".")[-1], None)
+
+        if not language_parser:
+            return []
+
+        group_ids = list(
+            Group.objects.filter(
+                first_seen__gte=datetime.now(UTC) - timedelta(days=90),
+                last_seen__gte=datetime.now(UTC) - timedelta(days=14),
+                status=GroupStatus.UNRESOLVED,
+                project__in=projects,
+            )
+            .order_by("-times_seen")
+            .values_list("id", flat=True)
+        )[:OPEN_PR_MAX_RECENT_ISSUES]
+        project_ids = [p.id for p in projects]
+
+        multi_if = language_parser.generate_multi_if(function_names)
+
+        # fetch the count of events for each group_id
+        subquery = (
+            Query(Entity("events"))
+            .set_select(
+                [
+                    Column("title"),
+                    Column("culprit"),
+                    Column("group_id"),
+                    Function("count", [], "event_count"),
+                    Function(
+                        "multiIf",
+                        multi_if,
+                        "function_name",
+                    ),
+                ]
+            )
+            .set_groupby(
+                [
+                    Column("title"),
+                    Column("culprit"),
+                    Column("group_id"),
+                    Column("exception_frames.function"),
+                ]
+            )
+            .set_where(
+                [
+                    Condition(Column("project_id"), Op.IN, project_ids),
+                    Condition(Column("group_id"), Op.IN, group_ids),
+                    Condition(Column("timestamp"), Op.GTE, datetime.now() - timedelta(days=14)),
+                    Condition(Column("timestamp"), Op.LT, datetime.now()),
+                    # NOTE: ideally this would follow suspect commit logic
+                    BooleanCondition(
+                        BooleanOp.OR,
+                        [
+                            BooleanCondition(
+                                BooleanOp.AND,
+                                [
+                                    Condition(
+                                        Function(
+                                            "arrayElement",
+                                            (Column("exception_frames.filename"), i),
+                                        ),
+                                        Op.IN,
+                                        sentry_filenames,
+                                    ),
+                                    language_parser.generate_function_name_conditions(
+                                        function_names, i
+                                    ),
+                                ],
+                            )
+                            for i in range(-STACKFRAME_COUNT, 0)  # first n frames
+                        ],
+                    ),
+                    Condition(Function("notHandled", []), Op.EQ, 1),
+                ]
+            )
+            .set_orderby([OrderBy(Column("event_count"), Direction.DESC)])
+        )
+
+        # filter on the subquery to squash group_ids with the same title and culprit
+        # return the group_id with the greatest count of events
+        query = (
+            Query(subquery)
+            .set_select(
+                [
+                    Column("function_name"),
+                    Function(
+                        "arrayElement",
+                        (Function("groupArray", [Column("group_id")]), 1),
+                        "group_id",
+                    ),
+                    Function(
+                        "arrayElement",
+                        (Function("groupArray", [Column("event_count")]), 1),
+                        "event_count",
+                    ),
+                ]
+            )
+            .set_groupby(
+                [
+                    Column("title"),
+                    Column("culprit"),
+                    Column("function_name"),
+                ]
+            )
+            .set_orderby([OrderBy(Column("event_count"), Direction.DESC)])
+            .set_limit(5)
+        )
+
+        request = SnubaRequest(
+            dataset=Dataset.Events.value,
+            app_id="default",
+            tenant_ids={"organization_id": projects[0].organization_id},
+            query=query,
+        )
+
+        try:
+            referrer = self.commit_context_referrers.pr_comment_bot.value
+            return raw_snql_query(request, referrer=referrer)["data"]
+        except Exception:
+            logger.exception(
+                _open_pr_comment_log(
+                    integration_name=self.integration_name, suffix="snuba_query_error"
+                ),
+                extra={"query": request.to_dict()["query"]},
+            )
+            return []
+
+    def get_issue_table_contents(
+        self, issue_list: list[IssueCountForFile]
+    ) -> list[PullRequestIssue]:
+        group_id_to_info = {}
+        for issue in issue_list:
+            group_id = issue["group_id"]
+            group_id_to_info[group_id] = dict(filter(lambda k: k[0] != "group_id", issue.items()))
+
+        issues = Group.objects.filter(id__in=list(group_id_to_info.keys())).all()
+
+        pull_request_issues = [
+            PullRequestIssue(
+                title=issue.title,
+                subtitle=issue.culprit,
+                url=issue.get_absolute_url(),
+                affected_users=issue.count_users_seen(
+                    referrer=Referrer.TAGSTORE_GET_GROUPS_USER_COUNTS_OPEN_PR_COMMENT.value
+                ),
+                event_count=group_id_to_info[issue.id]["event_count"],
+                function_name=group_id_to_info[issue.id]["function_name"],
+            )
+            for issue in issues
+        ]
+        pull_request_issues.sort(key=lambda k: k.event_count or 0, reverse=True)
+
+        return pull_request_issues
 
 
 def run_pr_comment_workflow(integration_name: str, pullrequest_id: int, project_id: int) -> None:
