+++ tests/sentry/integrations/github/tasks/test_open_pr_comment.py
@@ -9,78 +9,26 @@ from sentry.constants import ObjectStatus
 from sentry.integrations.github.tasks.open_pr_comment import (
     format_issue_table,
     format_open_pr_comment,
-    get_issue_table_contents,
     get_pr_files,
-    get_projects_and_filenames_from_source_file,
-    get_top_5_issues_by_count_for_file,
     open_pr_comment_workflow,
     safe_for_comment,
 )
-from sentry.integrations.github.tasks.utils import PullRequestFile, PullRequestIssue
 from sentry.integrations.models.integration import Integration
-from sentry.integrations.source_code_management.constants import STACKFRAME_COUNT
+from sentry.integrations.source_code_management.commit_context import (
+    PullRequestFile,
+    PullRequestIssue,
+)
 from sentry.integrations.source_code_management.language_parsers import PATCH_PARSERS
-from sentry.models.group import Group, GroupStatus
+from sentry.models.group import Group
 from sentry.models.pullrequest import CommentType, PullRequest, PullRequestComment
 from sentry.shared_integrations.exceptions import ApiError
 from sentry.testutils.cases import IntegrationTestCase, TestCase
-from sentry.testutils.helpers.datetime import before_now
 from sentry.testutils.silo import assume_test_silo_mode_of
 from sentry.testutils.skips import requires_snuba
 from tests.sentry.integrations.github.tasks.test_pr_comment import GithubCommentTestCase
+from tests.sentry.integrations.source_code_management.test_commit_context import CreateEventTestCase
 
 pytestmark = [requires_snuba]
-
-
-class CreateEventTestCase(TestCase):
-    def _create_event(
-        self,
-        culprit=None,
-        timestamp=None,
-        filenames=None,
-        function_names=None,
-        project_id=None,
-        user_id=None,
-        handled=False,
-    ):
-        if culprit is None:
-            culprit = "issue0"
-        if timestamp is None:
-            timestamp = before_now(seconds=5).isoformat()
-        if filenames is None:
-            filenames = ["foo.py", "baz.py"]
-        if function_names is None:
-            function_names = ["hello", "world"]
-        if project_id is None:
-            project_id = self.project.id
-
-        assert len(function_names) == len(filenames)
-
-        frames = []
-        for i, filename in enumerate(filenames):
-            frames.append({"filename": filename, "function": function_names[i]})
-
-        return self.store_event(
-            data={
-                "message": "hello!",
-                "culprit": culprit,
-                "platform": "python",
-                "timestamp": timestamp,
-                "exception": {
-                    "values": [
-                        {
-                            "type": "Error",
-                            "stacktrace": {
-                                "frames": frames,
-                            },
-                            "mechanism": {"handled": handled, "type": "generic"},
-                        },
-                    ]
-                },
-                "user": {"id": user_id},
-            },
-            project_id=project_id,
-        )
 
 
 class TestSafeForComment(GithubCommentTestCase):
@@ -257,438 +205,6 @@ class TestGetFilenames(GithubCommentTestCase):
         assert pr_file.filename == data[0]["filename"]
         assert pr_file.patch == data[0]["patch"]
 
-    def test_get_projects_and_filenames_from_source_file(self):
-        projects = [self.create_project() for _ in range(4)]
-
-        source_stack_pairs = [
-            ("", "./"),
-            ("src/sentry", "sentry/"),
-            ("src/", ""),
-            ("src/sentry/", "sentry/"),
-        ]
-        for i, pair in enumerate(source_stack_pairs):
-            source_root, stack_root = pair
-            self.create_code_mapping(
-                project=projects[i],
-                repo=self.gh_repo,
-                source_root=source_root,
-                stack_root=stack_root,
-                default_branch="master",
-            )
-
-        # matching code mapping from a different org
-        other_org_code_mapping = self.create_code_mapping(
-            project=self.another_org_project,
-            repo=self.another_org_repo,
-            source_root="",
-            stack_root="./",
-        )
-        other_org_code_mapping.organization_id = self.another_organization.id
-        other_org_code_mapping.save()
-
-        source_stack_nonmatches = [
-            ("/src/sentry", "sentry"),
-            ("tests/", "tests/"),
-            ("app/", "static/app"),
-            ("tasks/integrations", "tasks"),  # random match in the middle of the string
-        ]
-        for source_root, stack_root in source_stack_nonmatches:
-            self.create_code_mapping(
-                project=self.create_project(),
-                repo=self.gh_repo,
-                source_root=source_root,
-                stack_root=stack_root,
-                default_branch="master",
-            )
-
-        filename = "src/sentry/tasks/integrations/github/open_pr_comment.py"
-        correct_filenames = [
-            "./src/sentry/tasks/integrations/github/open_pr_comment.py",
-            "sentry//tasks/integrations/github/open_pr_comment.py",
-            "sentry/tasks/integrations/github/open_pr_comment.py",
-        ]
-
-        project_list, sentry_filenames = get_projects_and_filenames_from_source_file(
-            self.organization.id, self.gh_repo.id, filename
-        )
-        assert project_list == set(projects)
-        assert sentry_filenames == set(correct_filenames)
-
-    def test_get_projects_and_filenames_from_source_file_filters_repo(self):
-        projects = [self.create_project() for _ in range(3)]
-
-        source_stack_pairs = [
-            ("src/sentry", "sentry/"),
-            ("src/", ""),
-            ("src/sentry/", "sentry/"),
-        ]
-        for i, pair in enumerate(source_stack_pairs):
-            source_root, stack_root = pair
-            self.create_code_mapping(
-                project=projects[i],
-                repo=self.gh_repo,
-                source_root=source_root,
-                stack_root=stack_root,
-                default_branch="master",
-            )
-
-        # other codemapping in different repo, will not match
-        project = self.create_project()
-        repo = self.create_repo(
-            name="getsentry/santry",
-            provider="integrations:github",
-            integration_id=self.integration.id,
-            project=project,
-            url="https://github.com/getsentry/santry",
-        )
-        self.create_code_mapping(
-            project=project,
-            repo=repo,
-            source_root="",
-            stack_root="./",
-            default_branch="master",
-        )
-
-        filename = "src/sentry/tasks/integrations/github/open_pr_comment.py"
-        correct_filenames = [
-            "sentry//tasks/integrations/github/open_pr_comment.py",
-            "sentry/tasks/integrations/github/open_pr_comment.py",
-        ]
-
-        project_list, sentry_filenames = get_projects_and_filenames_from_source_file(
-            self.organization.id, self.gh_repo.id, filename
-        )
-        assert project_list == set(projects)
-        assert sentry_filenames == set(correct_filenames)
-
-
-class TestGetCommentIssues(CreateEventTestCase):
-    def setUp(self):
-        self.group_id = [self._create_event(user_id=str(i)) for i in range(6)][0].group.id
-        self.another_org = self.create_organization()
-        self.another_org_project = self.create_project(organization=self.another_org)
-
-    def test_simple(self):
-        group_id = [
-            self._create_event(function_names=["blue", "planet"], user_id=str(i)) for i in range(7)
-        ][0].group.id
-        top_5_issues = get_top_5_issues_by_count_for_file(
-            [self.project], ["baz.py"], ["world", "planet"]
-        )
-
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        function_names = [issue["function_name"] for issue in top_5_issues]
-        assert top_5_issue_ids == [group_id, self.group_id]
-        assert function_names == ["planet", "world"]
-
-    def test_javascript_simple(self):
-        # should match function name exactly or className.functionName
-        group_id_1 = [
-            self._create_event(
-                function_names=["other.planet", "component.blue"],
-                filenames=["baz.js", "foo.js"],
-                user_id=str(i),
-            )
-            for i in range(7)
-        ][0].group.id
-        group_id_2 = [
-            self._create_event(
-                function_names=["component.blue", "world"],
-                filenames=["foo.js", "baz.js"],
-                user_id=str(i),
-            )
-            for i in range(6)
-        ][0].group.id
-        top_5_issues = get_top_5_issues_by_count_for_file(
-            [self.project], ["baz.js"], ["world", "planet"]
-        )
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        function_names = [issue["function_name"] for issue in top_5_issues]
-        assert top_5_issue_ids == [group_id_1, group_id_2]
-        assert function_names == ["other.planet", "world"]
-
-    def test_php_simple(self):
-        # should match function name exactly or namespace::functionName
-        group_id_1 = [
-            self._create_event(
-                function_names=["namespace/other/test::planet", "test/component::blue"],
-                filenames=["baz.php", "foo.php"],
-                user_id=str(i),
-            )
-            for i in range(7)
-        ][0].group.id
-        group_id_2 = [
-            self._create_event(
-                function_names=["test/component::blue", "world"],
-                filenames=["foo.php", "baz.php"],
-                user_id=str(i),
-            )
-            for i in range(6)
-        ][0].group.id
-        top_5_issues = get_top_5_issues_by_count_for_file(
-            [self.project], ["baz.php"], ["world", "planet"]
-        )
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        function_names = [issue["function_name"] for issue in top_5_issues]
-        assert top_5_issue_ids == [group_id_1, group_id_2]
-        assert function_names == ["namespace/other/test::planet", "world"]
-
-    def test_ruby_simple(self):
-        # should match function name exactly or class.functionName
-        group_id_1 = [
-            self._create_event(
-                function_names=["test.planet", "test/component.blue"],
-                filenames=["baz.rb", "foo.rb"],
-                user_id=str(i),
-            )
-            for i in range(7)
-        ][0].group.id
-        group_id_2 = [
-            self._create_event(
-                function_names=["test/component.blue", "world"],
-                filenames=["foo.rb", "baz.rb"],
-                user_id=str(i),
-            )
-            for i in range(6)
-        ][0].group.id
-        top_5_issues = get_top_5_issues_by_count_for_file(
-            [self.project], ["baz.rb"], ["world", "planet"]
-        )
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        function_names = [issue["function_name"] for issue in top_5_issues]
-        assert top_5_issue_ids == [group_id_1, group_id_2]
-        assert function_names == ["test.planet", "world"]
-
-    def test_filters_resolved_issue(self):
-        group = Group.objects.all()[0]
-        group.resolved_at = timezone.now()
-        group.status = GroupStatus.RESOLVED
-        group.substatus = None
-        group.save()
-
-        top_5_issues = get_top_5_issues_by_count_for_file([self.project], ["baz.py"], ["world"])
-        assert len(top_5_issues) == 0
-
-    def test_filters_handled_issue(self):
-        group_id = self._create_event(filenames=["bar.py", "baz.py"], handled=True).group.id
-
-        top_5_issues = get_top_5_issues_by_count_for_file([self.project], ["baz.py"], ["world"])
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        assert group_id != self.group_id
-        assert top_5_issue_ids == [self.group_id]
-
-    def test_project_group_id_mismatch(self):
-        # we fetch all group_ids that belong to the projects passed into the function
-        self._create_event(project_id=self.another_org_project.id)
-
-        top_5_issues = get_top_5_issues_by_count_for_file([self.project], ["baz.py"], ["world"])
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        assert top_5_issue_ids == [self.group_id]
-
-    def test_filename_mismatch(self):
-        group_id = self._create_event(
-            filenames=["foo.py", "bar.py"],
-        ).group.id
-
-        top_5_issues = get_top_5_issues_by_count_for_file([self.project], ["baz.py"], ["world"])
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        assert group_id != self.group_id
-        assert top_5_issue_ids == [self.group_id]
-
-    def test_function_name_mismatch(self):
-        group_id = self._create_event(
-            function_names=["world", "hello"],
-        ).group.id
-
-        top_5_issues = get_top_5_issues_by_count_for_file([self.project], ["baz.py"], ["world"])
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        assert group_id != self.group_id
-        assert top_5_issue_ids == [self.group_id]
-
-    def test_not_first_frame(self):
-        group_id = self._create_event(
-            function_names=["world", "hello"], filenames=["baz.py", "bar.py"], culprit="hi"
-        ).group.id
-
-        top_5_issues = get_top_5_issues_by_count_for_file([self.project], ["baz.py"], ["world"])
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        function_names = [issue["function_name"] for issue in top_5_issues]
-        assert group_id != self.group_id
-        assert top_5_issue_ids == [self.group_id, group_id]
-        assert function_names == ["world", "world"]
-
-    def test_not_within_frame_limit(self):
-        function_names = ["world"] + ["a" for _ in range(STACKFRAME_COUNT)]
-        filenames = ["baz.py"] + ["foo.py" for _ in range(STACKFRAME_COUNT)]
-        group_id = self._create_event(function_names=function_names, filenames=filenames).group.id
-
-        top_5_issues = get_top_5_issues_by_count_for_file([self.project], ["baz.py"], ["world"])
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        assert group_id != self.group_id
-        assert top_5_issue_ids == [self.group_id]
-
-    def test_event_too_old(self):
-        group_id = self._create_event(
-            timestamp=before_now(days=15).isoformat(), filenames=["bar.py", "baz.py"]
-        ).group.id
-
-        top_5_issues = get_top_5_issues_by_count_for_file([self.project], ["baz.py"], ["world"])
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        assert group_id != self.group_id
-        assert top_5_issue_ids == [self.group_id]
-
-    def test_squashes_same_title_culprit_issues(self):
-        # both of these have the same title and culprit,
-        # so "squash" them and return the one with greater number of events
-        [
-            self._create_event(
-                filenames=["base.py", "baz.py"],
-                function_names=["wonderful", "world"],
-                user_id=str(i),
-                handled=False,
-            )
-            for i in range(3)
-        ]
-        group_id = [
-            self._create_event(
-                filenames=["bar.py", "baz.py"],
-                function_names=["blue", "planet"],
-                user_id=str(i),
-                handled=False,
-            )
-            for i in range(5)
-        ][0].group_id
-
-        top_5_issues = get_top_5_issues_by_count_for_file(
-            [self.project], ["baz.py"], ["world", "planet"]
-        )
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        function_names = [issue["function_name"] for issue in top_5_issues]
-
-        assert top_5_issue_ids == [self.group_id, group_id]
-        assert function_names == ["world", "planet"]
-
-    def test_fetches_top_five_issues(self):
-        group_id_1 = [
-            self._create_event(
-                filenames=["bar.py", "baz.py"],
-                function_names=["blue", "planet"],
-                user_id=str(i),
-                handled=False,
-            )
-            for i in range(5)
-        ][0].group.id
-        [
-            self._create_event(
-                filenames=["hello.py", "baz.py"],
-                function_names=["green", "planet"],
-                user_id=str(i),
-                handled=True,
-            )
-            for i in range(4)
-        ]
-        group_id_3 = [
-            self._create_event(
-                filenames=["base.py", "baz.py"],
-                function_names=["wonderful", "world"],
-                user_id=str(i),
-                handled=False,
-                culprit="hi",
-            )
-            for i in range(3)
-        ][0].group.id
-        [
-            self._create_event(
-                filenames=["nom.py", "baz.py"],
-                function_names=["jurassic", "world"],
-                user_id=str(i),
-                handled=True,
-            )
-            for i in range(2)
-        ]
-        # 6th issue
-        self._create_event(
-            filenames=["nan.py", "baz.py"], function_names=["my_own", "world"], handled=True
-        )
-        # unrelated issue with same stack trace in different project
-        self._create_event(project_id=self.another_org_project.id)
-
-        top_5_issues = get_top_5_issues_by_count_for_file(
-            [self.project], ["baz.py"], ["world", "planet"]
-        )
-        top_5_issue_ids = [issue["group_id"] for issue in top_5_issues]
-        function_names = [issue["function_name"] for issue in top_5_issues]
-
-        # filters handled issues
-        assert top_5_issue_ids == [self.group_id, group_id_1, group_id_3]
-        assert function_names == ["world", "planet", "world"]
-
-    def test_get_issue_table_contents(self):
-        group_id_1 = [
-            self._create_event(
-                culprit="issue1",
-                filenames=["bar.py", "baz.py"],
-                function_names=["blue", "planet"],
-                user_id=str(i),
-                handled=False,
-            )
-            for i in range(5)
-        ][0].group.id
-        group_id_2 = [
-            self._create_event(
-                culprit="issue2",
-                filenames=["hello.py", "baz.py"],
-                function_names=["green", "planet"],
-                user_id=str(i),
-                handled=False,
-            )
-            for i in range(4)
-        ][0].group.id
-        group_id_3 = [
-            self._create_event(
-                culprit="issue3",
-                filenames=["base.py", "baz.py"],
-                function_names=["wonderful", "world"],
-                user_id=str(i),
-                handled=False,
-            )
-            for i in range(3)
-        ][0].group.id
-        group_id_4 = [
-            self._create_event(
-                culprit="issue4",
-                filenames=["nom.py", "baz.py"],
-                function_names=["jurassic", "world"],
-                user_id=str(i),
-                handled=False,
-            )
-            for i in range(2)
-        ][0].group.id
-
-        top_5_issues = get_top_5_issues_by_count_for_file(
-            [self.project], ["baz.py"], ["world", "planet"]
-        )
-        affected_users = [6, 5, 4, 3, 2]
-        event_count = [issue["event_count"] for issue in top_5_issues]
-        function_names = [issue["function_name"] for issue in top_5_issues]
-
-        comment_table_contents = get_issue_table_contents(top_5_issues)
-        group_ids = [self.group_id, group_id_1, group_id_2, group_id_3, group_id_4]
-
-        for i in range(5):
-            subtitle = "issue" + str(i)
-            assert (
-                PullRequestIssue(
-                    title="Error",
-                    subtitle=subtitle,
-                    url=f"http://testserver/organizations/{self.organization.slug}/issues/{group_ids[i]}/",
-                    affected_users=affected_users[i],
-                    event_count=event_count[i],
-                    function_name=function_names[i],
-                )
-                in comment_table_contents
-            )
-
 
 class TestFormatComment(TestCase):
     def setUp(self):
@@ -822,12 +338,14 @@ Your pull request is modifying functions with the following pre-existing issues:
 
 @patch("sentry.integrations.github.tasks.open_pr_comment.get_pr_files")
 @patch(
-    "sentry.integrations.github.tasks.open_pr_comment.get_projects_and_filenames_from_source_file"
+    "sentry.integrations.github.integration.GitHubIntegration.get_projects_and_filenames_from_source_file"
 )
 @patch(
     "sentry.integrations.source_code_management.language_parsers.PythonParser.extract_functions_from_patch"
 )
-@patch("sentry.integrations.github.tasks.open_pr_comment.get_top_5_issues_by_count_for_file")
+@patch(
+    "sentry.integrations.github.integration.GitHubIntegration.get_top_5_issues_by_count_for_file"
+)
 @patch("sentry.integrations.github.tasks.open_pr_comment.safe_for_comment")
 @patch("sentry.integrations.source_code_management.commit_context.metrics")
 class TestOpenPRCommentWorkflow(IntegrationTestCase, CreateEventTestCase):
@@ -1034,7 +552,7 @@ class TestOpenPRCommentWorkflow(IntegrationTestCase, CreateEventTestCase):
         assert not mock_analytics.called
 
     @patch("sentry.analytics.record")
-    @patch("sentry.integrations.github.tasks.open_pr_comment.metrics")
+    @patch("sentry.integrations.github.integration.metrics")
     @responses.activate
     def test_comment_workflow_api_error(
         self,
